Add armv5te release build for Marvell Kirkwood NAS boxes (Iomega ix2-dl)
Debian Trixie's armel port targets exactly this: ARMv5TE soft-float,
no hardware FPU or 64-bit atomics. libpcap is still statically linked
(built from source with the arm-linux-gnueabi cross toolchain, same
recipe already used for the musl targets), so the only extra runtime
dependency versus the other Linux builds is glibc itself.

Three AtomicU64 usages had to become AtomicU32, all without losing
real range:
- event.rs / EventHandler::tick_rate_ms: clamped to 100..=5000 at
  every write site, same as syswatch's equivalent.
- collectors/packets/mod.rs / CaptureStats: received/dropped/rate_pps
  are sourced from libpcap's pcap_stats(), whose C struct fields
  (ps_recv/ps_drop/ps_ifdrop) are `u_int` -- 32 bits on every
  platform pcap supports, not just this one.
- remote/queue.rs / SnapshotQueue::dropped: counts overflow drops of
  whole snapshots (one push per agent tick), so wrapping would take
  centuries even under sustained overload.

All three keep their existing u64 public API (get/set take or return
u64); only the internal atomic storage narrowed.

+++ src/event.rs
@@ -1,6 +1,6 @@
 use anyhow::Result;
 use crossterm::event::{self, Event, KeyEvent, KeyEventKind, MouseEvent};
-use std::sync::atomic::{AtomicU64, Ordering};
+use std::sync::atomic::{AtomicU32, Ordering};
 use std::sync::Arc;
 use std::time::Duration;
 use tokio::sync::mpsc;
@@ -16,13 +16,16 @@ pub struct EventHandler {
     /// Polling interval for the background thread, in milliseconds.
     /// Shared with the thread so settings changes apply within one
     /// poll cycle without requiring a restart.
-    tick_rate_ms: Arc<AtomicU64>,
+    // u32, not u64: the value is clamped to 100..=5000 right below, which
+    // fits comfortably, and armv5te (old Kirkwood NAS boxes) has no native
+    // 64-bit atomics.
+    tick_rate_ms: Arc<AtomicU32>,
 }
 
 impl EventHandler {
     pub fn start(tick_rate_ms: u64) -> Self {
         let (tx, rx) = mpsc::unbounded_channel();
-        let tick_rate_ms = Arc::new(AtomicU64::new(tick_rate_ms.clamp(100, 5000)));
+        let tick_rate_ms = Arc::new(AtomicU32::new(tick_rate_ms.clamp(100, 5000) as u32));
         let tick_rate_thread = Arc::clone(&tick_rate_ms);
 
         // Use a dedicated OS thread instead of tokio::spawn, since
@@ -33,7 +36,7 @@ impl EventHandler {
                 // Re-read the tick rate each iteration so the settings popup
                 // can change refresh_rate_ms at runtime and have it take
                 // effect on the next poll without a restart.
-                let dur = Duration::from_millis(tick_rate_thread.load(Ordering::Relaxed));
+                let dur = Duration::from_millis(tick_rate_thread.load(Ordering::Relaxed) as u64);
                 if event::poll(dur).unwrap_or(false) {
                     match event::read() {
                         Ok(Event::Key(key)) if key.kind == KeyEventKind::Press => {
@@ -62,7 +65,7 @@ impl EventHandler {
     /// iteration without tracking previous state.
     pub fn set_tick_rate(&self, ms: u64) {
         self.tick_rate_ms
-            .store(ms.clamp(100, 5000), Ordering::Relaxed);
+            .store(ms.clamp(100, 5000) as u32, Ordering::Relaxed);
     }
 
     pub async fn next(&mut self) -> Result<AppEvent> {
